+++ application/views/admin/system/system.php
@@ -40,8 +40,6 @@ defined('BASEPATH') OR exit('No direct script access allowed');
             <div class="large-12 columns">
                 <input type="submit" value="<?= $lang->line('submit') ?>" class="button small round">
             </div>
-
-            <?php //var_dump($options) ?>
         </fieldset>
     </form>
 </div>