+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Promise First Version Release before 2014-2
 
 ------------
 
-* ###### 2013-8-28 System Options Doing
+* ! 2013-9-2 System Languages Doing
+* ! 2013-8-28 System Options Doing
 * 2013-9-1 Location using Google Maps Done
 * 2013-8-28 Dashboard Style Done
 * 2013-8-27 Cities Management Done